requirements.txt
colour-science==0.3.16
pyserial==3.5
